requirements.txt
selenium==4.16.0
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0
webdriver-manager==4.0.1
python-dotenv==1.0.0
allure-pytest==2.13.2
pytest-rerunfailures==12.0
